// src/dorc-web/src/apis/dorc-api/models/ApiApiModel.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @enum {string}
 */
export enum ApiEndpointResolutionStatus {
    NoTokens = 'NoTokens',
    Resolved = 'Resolved',
    PartiallyResolved = 'PartiallyResolved'
}

/**
 * @export
 * @interface ApiApiModel
 */
export interface ApiApiModel {
    /**
     * @type {number}
     * @memberof ApiApiModel
     */
    Id?: number;
    /**
     * @type {number}
     * @memberof ApiApiModel
     */
    EnvironmentId?: number;
    /**
     * @type {string}
     * @memberof ApiApiModel
     */
    EnvironmentName?: string | null;
    /**
     * @type {string}
     * @memberof ApiApiModel
     */
    Name?: string | null;
    /**
     * @type {string}
     * @memberof ApiApiModel
     */
    Endpoint?: string | null;
    /**
     * @type {string}
     * @memberof ApiApiModel
     */
    EndpointResolved?: string | null;
    /**
     * @type {ApiEndpointResolutionStatus}
     * @memberof ApiApiModel
     */
    ResolutionStatus?: ApiEndpointResolutionStatus;
    /**
     * @type {string}
     * @memberof ApiApiModel
     */
    UnresolvedTokens?: string | null;
    /**
     * @type {string}
     * @memberof ApiApiModel
     */
    Description?: string | null;
    /**
     * @type {string}
     * @memberof ApiApiModel
     */
    Type?: string | null;
    /**
     * @type {string}
     * @memberof ApiApiModel
     */
    AuthType?: string | null;
    /**
     * @type {string}
     * @memberof ApiApiModel
     */
    HealthCheckPath?: string | null;
    /**
     * @type {number}
     * @memberof ApiApiModel
     */
    OwnerProjectId?: number | null;
    /**
     * @type {string}
     * @memberof ApiApiModel
     */
    OwnerProjectName?: string | null;
    /**
     * @type {string}
     * @memberof ApiApiModel
     */
    Tags?: string | null;
    /**
     * @type {boolean}
     * @memberof ApiApiModel
     */
    UserEditable?: boolean;
}